Update generalExpLog.tex
Minor changes: The derivative rule works for a>0, not just a>1. Added an algebraic step for clarity in the derivation of the rule. Also, while I didn't make this change above, I would consider calling on log rules rather than a base conversion formula to reinforce the log identities to the students:  a^x = e^{\ln(a^x)} = e^{x\ln(a)}. This appeals to something that they'll use again in the course, rather than an esoteric identity.